--- src/SchedulerDriver/Request.java
@@ -1,20 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package SchedulerDriver;
-
-/**
- *
- * @author Hayden
- */
-
-public class Request {
-    public String agent;
-    public reqType rORc;
-    public seatType type;
-    public int seatNum;
-    public int arrivalTime;
-    
-}
